+++ annotations/nocturne.exe/pseudocode/src/core/scat.cpp/CScat_updateAiming_FUN_004fd450.cpp
@@ -76,8 +76,8 @@ void __cdecl core_scat_cpp_CScat_updateAiming_FUN_004fd450(CScat *this_ptr,float
     for (local_20 = is_holstered; local_20 < g_CDemonSet_PTR_005be368->actor_count;
         local_20 = local_20 + 1) {
       target = *(CDemonActor **)((int)g_CDemonSet_PTR_005be368->actors + iVar3);
-      local_18 = (float)core_scat_cpp_CScat_scoreAimTarget_FUN_004fda20
-                                  (this_ptr,target,(uint)(target == this_ptr->aim_target));
+      local_18 = core_scat_cpp_CScat_scoreAimTarget_FUN_004fda20
+                           (this_ptr,target,(uint)(target == this_ptr->aim_target));
       if ((0.0 <= local_18) && (local_18 < local_24)) {
         local_28 = target;
         local_24 = local_18;
